Refactor simulation into package and add tqdm
